Redundancia en redux de users, movido a modulos de user
