+++ client/src/redux/actions/actions.js
@@ -1,59 +1,14 @@
 import axios from "axios";
-import { GET_ARTICLE_TAG } from "./actionsArticles";
 
 const { REACT_APP_URL_API } = process.env;
 
 export const GET_ALL_CATEGORIES = "GET ALL CATEGORIES";
-export const GET_USERS = "GET USERS";
-export const SET_USER_ID = "SET_USER_ID";
-export const SET_USER_ROLES = "SET_USER_ROLES";
 export const GET_ALL_CAT_SUB = "GET_ALL_CAT_SUB";
 export const GET_CATEGORY = "GET_CATEGORY";
 export const GET_SUB_CATEGORY = "GET_SUB_CATEGORY";
-export const GET_USERS_BY_ROLES = "GET_USERS_BY_ROLES";
+
+//////////////////////////////////////
 export const GET_ADMINS = "GET_ADMINS";
-
-
-export const getAllUsers = () => {
-  return async (dispatch) => {
-    try {
-      const response = await axios.get(`${REACT_APP_URL_API}/users`);
-      dispatch({
-        type: GET_USERS,
-        payload: response.data,
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-};
-
-export const findOrCreateUser = (user, token) => async (dispatch) => {
-  try {
-    const userPost = {
-      user_id_A0: user.sub,
-      user_name: user.name,
-      user_email: user.email,
-      user_img_profile: user.picture,
-      inst_id: [],
-    };
-    const headers = {
-      Authorization: `Bearer ${token}`,
-    };
-    const response = await axios.post(`${REACT_APP_URL_API}/users`, userPost, {
-      headers,
-    });
-    if (response.data.user_id) {
-      dispatch(setUserID(response.data.user_id));
-    }
-    if (response.data.roles) {
-      dispatch(setUserRoles(response.data.roles));
-    }
-  } catch (err) {
-    console.log(err);
-    return;
-  }
-};
 
 export const setCategory = (body, token) => async (dispatch) => {
   try {
@@ -209,13 +164,6 @@ export const createNewSubCategory =
     }
   };
 
-export const setUserID = (payload) => {
-  return { type: SET_USER_ID, payload: payload };
-};
-
-export const setUserRoles = (payload) => {
-  return { type: SET_USER_ROLES, payload: payload };
-};
 
 export const createPost = (newPost, token) => async (dispatch) => {
   try {
@@ -293,19 +241,7 @@ export const getCategory = (id) => {
   };
 };
 
-export const getUsersByRoles = (rol) => {
-  return async (dispatch) => {
-    try {
-      const response = await axios.get(`${REACT_APP_URL_API}/users?rol=${rol}`);
-      dispatch({
-        type: GET_USERS_BY_ROLES,
-        payload: response.data,
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-};
+
 
 export const getSubCategory = (id) => {
   return async (dispatch) => {
