+++ client/src/views/Art_Detail/Art_Detail.jsx
@@ -6,10 +6,11 @@ import { useSelector, useDispatch } from "react-redux";
 import {
   getArticleDetail,
   clearDetail,
-} from '../../redux/actions/actionsArticles'
+} from '../../redux/actions/actionsArticles';
 import {
-
-  getAllUsers,
+  getAllUsers
+} from '../../redux/actions/usersActions';
+import {
   deletePost,
   getAllCatSub,
 } from "../../redux/actions/actions";
@@ -90,9 +91,9 @@ const Art_Detail = () => {
 
   const dispatch = useDispatch();
   const articlesDetail = useSelector((state) => state.articlesReducer.articlesDetail); // Nueva forma de acceder al estado por combineReducer
-  const users = useSelector((state) => state.rootReducer.users); // Nueva forma de acceder al estado por combineReducer
-  const user_id = useSelector((state) => state.rootReducer.user_id); // Nueva forma de acceder al estado por combineReducer
-	const user_roles = useSelector((state) => state.rootReducer.user_roles);
+  const users = useSelector((state) => state.usersReducer.users); // Nueva forma de acceder al estado por combineReducer
+  const user_id = useSelector((state) => state.usersReducer.user_id); // Nueva forma de acceder al estado por combineReducer
+	const user_roles = useSelector((state) => state.usersReducer.user_roles);
   const [idUser, setIdUser] = useState([]);
   const subcategories = useSelector((state) => state.rootReducer.cat_sub?.sub_cats);
   // const [section, setSection] = useState([]);
